fix: Tool workflow trigger execution error
